honksquad: fix: add validation, feedback, and extract healing logic in surgery system
Add popups for silent failure paths (already draped, patient stood up,
bedsheet missing, busy, invalid procedure, no organs). Validate organ
removal is at correct step to prevent step-skipping exploits. Check
container insert/remove return values. Add default case warning for
unknown surgery effects. Extract healing logic to SurgerySystem.Healing.cs
partial class to keep main file under 400 lines.
