+++ Content.Server/RussStation/Surgery/SurgerySystem.cs
@@ -1,17 +1,12 @@
-using System.Linq;
 using Content.Shared.Alert;
 using Content.Shared.Body;
 using Content.Shared.Body.Systems;
-using Content.Shared.Damage;
-using Content.Shared.Damage.Components;
 using Content.Shared.Damage.Systems;
 using Content.Shared.DoAfter;
-using Content.Shared.FixedPoint;
 using Content.Shared.Interaction;
 using Content.Shared.Popups;
 using Content.Shared.RussStation.Surgery;
 using Content.Shared.RussStation.Surgery.Components;
-using Content.Shared.RussStation.Surgery.Effects;
 using Content.Shared.RussStation.Surgery.Systems;
 using Content.Shared.Standing;
 using Content.Shared.Tag;
@@ -164,14 +159,23 @@ public sealed partial class SurgerySystem : SharedSurgerySystem
 
         // Validate: patient must still be down and not already draped
         if (HasComp<SurgeryDrapedComponent>(target.Value))
+        {
+            _popup.PopupEntity(Loc.GetString("surgery-already-draped"), target.Value, surgeon);
             return;
+        }
 
         if (!_standing.IsDown(target.Value))
+        {
+            _popup.PopupEntity(Loc.GetString("surgery-patient-not-down"), target.Value, surgeon);
             return;
+        }
 
         // Validate: bedsheet must still exist and have the Bedsheet tag
         if (!Exists(bedsheet.Value) || !_tags.HasTag(bedsheet.Value, "Bedsheet"))
+        {
+            _popup.PopupEntity(Loc.GetString("surgery-bedsheet-missing"), target.Value, surgeon);
             return;
+        }
 
         // Now drape the patient and take the bedsheet
         var draped = EnsureComp<SurgeryDrapedComponent>(target.Value);
@@ -179,7 +183,11 @@ public sealed partial class SurgerySystem : SharedSurgerySystem
         Dirty(target.Value, draped);
 
         var drapeContainer = _container.EnsureContainer<Container>(target.Value, "surgery_drape");
-        _container.Insert(bedsheet.Value, drapeContainer);
+        if (!_container.Insert(bedsheet.Value, drapeContainer))
+        {
+            Log.Warning($"Failed to insert bedsheet {ToPrettyString(bedsheet.Value)} into surgery drape container on {ToPrettyString(target.Value)}");
+            return;
+        }
 
         _popup.PopupEntity(Loc.GetString("surgery-drape-patient", ("target", target.Value)), target.Value);
 
@@ -198,10 +206,16 @@ public sealed partial class SurgerySystem : SharedSurgerySystem
     {
         if (active.ProcedureId == null ||
             !ProtoManager.TryIndex<SurgeryProcedurePrototype>(active.ProcedureId.Value, out var proto))
+        {
+            _popup.PopupEntity(Loc.GetString("surgery-procedure-invalid"), patient, surgeon);
             return;
+        }
 
         if (active.CurrentStep >= proto.Steps.Count)
+        {
+            _popup.PopupEntity(Loc.GetString("surgery-procedure-complete"), patient, surgeon);
             return;
+        }
 
         var currentStep = proto.Steps[active.CurrentStep];
 
@@ -246,7 +260,8 @@ public sealed partial class SurgerySystem : SharedSurgerySystem
             BreakOnHandChange = true,
         };
 
-        _doAfter.TryStartDoAfter(doAfterArgs);
+        if (!_doAfter.TryStartDoAfter(doAfterArgs))
+            _popup.PopupEntity(Loc.GetString("surgery-busy"), patient, surgeon);
     }
 
     private void StartCauteryClose(EntityUid surgeon, EntityUid patient, EntityUid tool)
@@ -267,7 +282,8 @@ public sealed partial class SurgerySystem : SharedSurgerySystem
             BreakOnHandChange = true,
         };
 
-        _doAfter.TryStartDoAfter(doAfterArgs);
+        if (!_doAfter.TryStartDoAfter(doAfterArgs))
+            _popup.PopupEntity(Loc.GetString("surgery-busy"), patient, surgeon);
     }
 
     private void OnStepDoAfter(Entity<ActiveSurgeryComponent> ent, ref SurgeryStepDoAfterEvent args)
@@ -281,10 +297,16 @@ public sealed partial class SurgerySystem : SharedSurgerySystem
 
         if (active.ProcedureId == null ||
             !ProtoManager.TryIndex<SurgeryProcedurePrototype>(active.ProcedureId.Value, out var proto))
+        {
+            Log.Warning($"Surgery step DoAfter completed but procedure {active.ProcedureId} is invalid on {ToPrettyString(patient)}");
             return;
+        }
 
         if (active.CurrentStep >= proto.Steps.Count)
+        {
+            _popup.PopupEntity(Loc.GetString("surgery-procedure-complete"), patient);
             return;
+        }
 
         var step = proto.Steps[active.CurrentStep];
 
@@ -333,116 +355,4 @@ public sealed partial class SurgerySystem : SharedSurgerySystem
         ApplyCauteryClose(ent.Owner, args.User);
     }
 
-    private void ApplyStepEffects(EntityUid patient, SurgeryStep step)
-    {
-        if (step.Damage != null)
-            _damageable.TryChangeDamage(patient, step.Damage);
-
-        if (step.Healing != null)
-        {
-            if (step.HealingTotal > 0 && TryComp<DamageableComponent>(patient, out var damageable))
-            {
-                // Distribute a fixed healing budget proportionally across actual damage
-                var budget = FixedPoint2.New(step.HealingTotal);
-                var currentDamage = _damageable.GetPositiveDamage((patient, damageable));
-                var totalDamage = FixedPoint2.Zero;
-
-                foreach (var type in step.Healing.DamageDict.Keys)
-                {
-                    if (currentDamage.DamageDict.TryGetValue(type, out var current))
-                        totalDamage += current;
-                }
-
-                if (totalDamage > 0)
-                {
-                    var healSpec = new DamageSpecifier();
-                    foreach (var type in step.Healing.DamageDict.Keys)
-                    {
-                        if (currentDamage.DamageDict.TryGetValue(type, out var current) && current > 0)
-                        {
-                            var share = budget * current / totalDamage;
-                            healSpec.DamageDict[type] = -share;
-                        }
-                    }
-
-                    _damageable.TryChangeDamage(patient, healSpec, true);
-                }
-            }
-            else
-            {
-                // No budget cap: heal each type independently
-                var negated = new DamageSpecifier(step.Healing);
-                foreach (var key in negated.DamageDict.Keys.ToList())
-                {
-                    negated.DamageDict[key] = -negated.DamageDict[key];
-                }
-
-                _damageable.TryChangeDamage(patient, negated, true);
-            }
-        }
-
-        if (step.BleedModifier != 0)
-            _bloodstream.TryModifyBleedAmount((patient, null), step.BleedModifier);
-    }
-
-    private void ApplyCauteryClose(EntityUid patient, EntityUid? surgeon)
-    {
-        // Cautery burn damage
-        var damage = new DamageSpecifier();
-        damage.DamageDict.Add("Heat", FixedPoint2.New(2));
-        _damageable.TryChangeDamage(patient, damage);
-
-        // Stop all bleeding
-        _bloodstream.TryModifyBleedAmount((patient, null), -100f);
-
-        if (surgeon != null)
-            _popup.PopupEntity(Loc.GetString("surgery-step-cauterize", ("user", surgeon.Value), ("target", patient)), patient);
-
-        // Clean up
-        RemComp<ActiveSurgeryComponent>(patient);
-        RemComp<SurgeryDrapedComponent>(patient); // Triggers OnDrapedShutdown -> drops bedsheet
-    }
-
-    private void HandleEffect(EntityUid? surgeon, EntityUid patient, ISurgeryEffect effect)
-    {
-        switch (effect)
-        {
-            case HealDamageEffect heal:
-                if (heal.Healing != null)
-                {
-                    var negated = new DamageSpecifier(heal.Healing);
-                    foreach (var key in negated.DamageDict.Keys.ToList())
-                    {
-                        negated.DamageDict[key] = -negated.DamageDict[key];
-                    }
-
-                    _damageable.TryChangeDamage(patient, negated, true);
-                }
-
-                break;
-
-            case RemoveOrganEffect:
-                OpenOrganRemovalMenu(surgeon, patient);
-                break;
-        }
-    }
-
-    private bool StepCanStillHeal(EntityUid patient, SurgeryStep step)
-    {
-        if (step.Healing == null || step.HealingTotal <= 0)
-            return false;
-
-        if (!TryComp<DamageableComponent>(patient, out var damageable))
-            return false;
-
-        var currentDamage = _damageable.GetPositiveDamage((patient, damageable));
-
-        foreach (var type in step.Healing.DamageDict.Keys)
-        {
-            if (currentDamage.DamageDict.TryGetValue(type, out var amount) && amount > 0)
-                return true;
-        }
-
-        return false;
-    }
 }
